+++ src/arduino-CAN/library.properties
